pyproject.toml
[project]
name = "ci-tkf-data-logic-svc"
version = "0.1.0"
description = "TKF Data Logic Service"
readme = "README.md"
requires-python = ">=3.10,<4.0.0"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0", 
    "prometheus-fastapi-instrumentator==6.1.0",
    "email-validator==2.1.1",
    "requests==2.31.0",
    "psycopg2-binary==2.9.11",
    "python-dotenv==1.0.0",
    "cryptography (>=46.0.3,<47.0.0)",
    "neo4j (==6.0.3)",
    "atlas-provider-sqlalchemy (==0.4.1)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
]

[project.optional-dependencies]
dev = [
    "black==23.11.0", 
    "flake8==6.1.0",
    "httpx==0.24.1",
]

[tool.poetry]
packages = [{include = "server", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0.0"
fastapi = "0.104.1"
uvicorn = {extras = ["standard"], version = "0.24.0"}
prometheus-fastapi-instrumentator = "6.1.0"
email-validator = "2.1.1"
requests = "2.31.0"
psycopg2-binary = "2.9.11"
python-dotenv = "1.0.0"

[tool.poetry.group.dev.dependencies]
pytest-cov = "4.1.0"
black = "23.11.0"
flake8 = "6.1.0"
httpx = "0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["src"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503", "W293", "F401"]